+++ packages/generator-nick/generators/app/templates/src/events/index.js
@@ -1,0 +1,9 @@
+/**
+ * Point of contact for events.
+ * @module events
+ * @example import events from './events';
+ */
+
+import events from '@robgietema/nick/src/events';
+
+export default events;