+++ src/routes/form/form.js
@@ -1,0 +1,17 @@
+/**
+ * Form routes.
+ * @module routes/form/form
+ */
+
+export default [
+  {
+    op: 'post',
+    view: '/@schemaform-data',
+    permission: 'View',
+    handler: async (req, trx) => {
+      return {
+        json: {},
+      };
+    },
+  },
+];